Filters out posts that don't have platformItemId and adTargetingData.
I chose to filter them out in sendPosts as sendPosts was already
filtering for posts that were already in the archive.

The other place you might think to put the filter was in `applyFilters`
as invoked in `feedScanner.js`. However, this place is not correct,
because `applyFilters` doesn't mean what you think it does. It does not
mean "applying filters". It actually means "apply info on posts that
pass this filter". What kind of "info"? Something like
applyAdTargetingInfo and applyAdId (going off the top of my head).

So...applyFilters wasn't actually doing any filtering at all. Very
confusing. It should be renamed.