coreDataGrab shouldn't analyze elements' siblings
coreDataGrab exists to find the ad ID and token pertaining to an element by
recursively iterating through its keys -- and those of any
elements/arrays/objects that are values -- to find anything that seems like
an ad ID or client token. Previously, this would analyze the `return` key
which is something like the parent element. This meant that it would traverse
up the tree and back down to siblings, which led to mistakes.