BUG: fix how to test a Facebook post is sponsored (broken recently).
Since Facebook, by FTC law, must have an aria label with the word "Sponsored",
our strategy is to find all elements with a aria-labelledby attr, then lookup
the corresponding aria label for the word Sponsored.

This is also not computationally expensive, since on the test post element
I used, I found only 3 aria-labelledby elements that we need to go through.